requirements.txt
pytube~=15.0.0
requests~=2.25.1
python-dotenv~=1.0.0
python-telegram-bot~=12.7